Constructing aircraft from templates, bugfixes
Aircraft template files are now parsed and aircraft are now constructed
from templates. In fact, that is now the only way to construct an
aircraft. Fixed a bug causing the application to crash when no aircraft
where selected, a couple aircraft where removed upon leaving boundaries
and then the user tried to select/deselect an aircraft. Refactoring and
code improvements, some things were moved to a separate file.